+++ docs/sections/changelog.rst
@@ -11,7 +11,7 @@ Unreleased
 * Added: token instrospection endpoint support (RFC7662).
 * Added: request in password grant authenticate call.
 * Changed: dropping support for Django versions before 1.8.
-* Changed: pass scope, token and request to OIDC_IDTOKEN_PROCESSING_HOOK.
+* Changed: pass token and request to OIDC_IDTOKEN_PROCESSING_HOOK.
 
 0.6.0
 =====